move initialization code to own module
